Add config prop
This way, a config object that might otherwise be passed to
`Plotly.setPlotConfig()` or as the last argument to `Plotly.newPlot()`,
can be passed to the `Plotly` Vue component.  It is still possible to
pass individial config props that get merged into the resulting config
object, but using the new config prop allows for encapsulation of
configuration as well as reuse and may help keep things tidier.